requirements.txt
# ── Python / ML dependencies ──────────────────────────────────────────────────
numpy>=1.24
pandas>=2.0
scikit-learn>=1.3
tensorflow==2.15.0
nltk>=3.8
spacy>=3.6

# ── Google Gemini ─────────────────────────────────────────────────────────────
google-generativeai>=0.5.0

# ── Django Backend ────────────────────────────────────────────────────────────
django~=3.2.0
djangorestframework>=3.12
django-cors-headers>=3.13
python-dotenv>=1.0
gunicorn>=21.0
joblib>=1.3
djongo>=1.3.6
pymongo==3.12.1
